feat: move update room vote deadline to Vote Api
